sensors: Light: Make sure events can be handled before enabling sensors.
Discussed with Lin this afternoon and we thought
it best to try to have Listeners removed prior
to disable and prior to enable.

Proximity driver is also being changed to,
in almost all cases, preventing input events
from arriving before this code set up for
it's newly enabled light sensor.

Signed-off-by: Pete Delaney  <piet@cyngn.com>
Signed-off-by: Lin Ma        <lma@cyngn.com>
Signed-off-by: Chirayu Desai <chirayudesai1@gmail.com>
Signed-off-by: Erica Chang <echang@cyngn.com>
Change-Id: Ie41aea925be5a4b5bde625599f17a62412b26dff