sensors: switch timestamps to CLOCK_BOOTIME
The Android CDD requires that timestamps 'must be synchronized with the
elapsedRealtimeNano clock'. elapsedRealtimeNano() is CLOCK_BOOTIME.

Proximity, Light and Mag Kernel sensor drivers have been updated to
report the timestamp in CLOCK_BOOTTIME using the EV_SYN_TIME*
event codes.

getTimestamp() is updated to stay in-sync with the Kernel time source.

LETTUCE-779

[1] https://source.android.com/devices/sensors/hal-interface.html#sensors_event_t

Change-Id: I866b42884fc7bd2ccd01cee335047abceb5d0ab3