Update patch set 1
Patch Set 1: Code-Review-2

On lettuce the proximity sensor is interrupt driven and does not make use of this delay. See here in the kernel:
 - https://github.com/CyanogenMod/android_kernel_cyanogen_msm8916/blob/588050e859bfc61eb517f53148661525a673ee31/drivers/input/misc/ap3426_cm.c#L1582

The proximity sensor itself is responsive in my testing. Can you provide a clear test case where it's not?

Regarding your comment about adaptive display, I'm not sure I follow. If you're using a shared ALS/Prox driver, it should be keeping track of the sensor rates for each independently. So, fixing the prox sensor might not be the correct thing to do.

Patch-set: 1